Refactor statement creation
