+++ presto-main/src/main/java/com/facebook/presto/Session.java
@@ -17,6 +17,7 @@ import com.facebook.presto.client.ClientSession;
 import com.facebook.presto.execution.QueryId;
 import com.facebook.presto.metadata.SessionPropertyManager;
 import com.facebook.presto.spi.ConnectorSession;
+import com.facebook.presto.spi.PrestoException;
 import com.facebook.presto.spi.security.Identity;
 import com.facebook.presto.spi.type.TimeZoneKey;
 import com.facebook.presto.transaction.TransactionId;
@@ -34,6 +35,7 @@ import java.util.Map.Entry;
 import java.util.Optional;
 import java.util.TimeZone;
 
+import static com.facebook.presto.spi.StandardErrorCode.NOT_FOUND;
 import static com.google.common.base.MoreObjects.toStringHelper;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkState;
@@ -195,6 +197,15 @@ public final class Session
     public Map<String, String> getPreparedStatements()
     {
         return preparedStatements;
+    }
+
+    public String getPreparedStatement(String name)
+    {
+        String sqlString = preparedStatements.get(name);
+        if (sqlString == null) {
+            throw new PrestoException(NOT_FOUND, "Prepared statement not found: " + name);
+        }
+        return sqlString;
     }
 
     public Session withTransactionId(TransactionId transactionId)
